compiler warning

+++ formatgs.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: formatgs.c,v 1.25 2008/06/23 23:08:00 c_anthon Exp $
+ * $Id: formatgs.c,v 1.26 2008/06/24 16:32:00 Superfly_Jon Exp $
  */
 
 #include "config.h"
@@ -600,7 +600,7 @@ formatGS( const statcontext *psc, const int nMatchTo,
       
           for ( j = 0; j < 2; ++j ) 
 	  {
-		  float ci = 1.95996f * sqrt( af[ i ][ 1 ][ j ] / psc->nGames );
+		  float ci = 1.95996f * sqrtf( af[ i ][ 1 ][ j ] / psc->nGames );
 		  float max = af[i][0][j]+ci;
 		  float min = af[i][0][j]-ci;
 		  aasz[ j + 1 ] = g_strdup_printf( "%*.*f - %*.*f",
